* views added to measurement app
